Initial check in of Shark ORM @ v2.00.0, which replaces DBAccess v1.6.13, but remains API compatible.
In order for DBAccess to be open sourced, we have to rewrite large swathes of the project and remove several areas due to time contraints, this was so the project was completely different to the original.  These include cache modules and query hashing, also a more advanced event mechanisum.

However, because the API for DBAccess was already in the public domain we have retained it, and further more we have included a DBAccess.h file which will enable you to continue to use your existing implementation without modification.