Fix missing "Id" property in the class definition cache.
[https://github.com/sharksync/sharkorm/issues/18]

Introduced when adding the support for inheritance in SRKObjects, we
skipped looking all the way back to SRKObject, and deliberately
excluded it.  Therefore the Id object was never defined and forever
returned an UNDEFINED type.  Luckily, in most cases this never actually
caused an issue because we were always testing for a string type to
change the code path.