Re-wrote the transaction handler
Features/Improvements:

Transactions are now fully isolated by thread.
Transactions are failed if ANY error occurs, including a syntax error in a query.
Any modifications to an object, it’s relationships or any changes it makes using triggers or the trigger methods (entityWillUpdate, delete etc) for cascading actions will also be rolled back, not just the SQL table contents but the entire object stack will be as it was before you called the transaction.
You can manually fail a transaction using ‘SRKFailTransaction()’, for any applicaiton logic you wish to apply.

Signed-off-by: Adrian Herridge <ah@Adrians-MacBook-Pro.local>